+++ src/platform.h
@@ -15,6 +15,8 @@
 // I took what I needed for simpleomp from:
 // https://raw.githubusercontent.com/Tencent/ncnn/9e11dac7d17aae3c600d30de471b57478459a624/src/platform.h.in
 
+#pragma once
+
 #include <pthread.h>
 
 #define NCNN_EXPORT